requirements.txt
Django
djangorestframework
django-environ
django-cors-headers
Pillow
djangorestframework-simplejwt
drf-spectacular
django-filter
